+++ runner/impl/src/main/java/com/walmartlabs/concord/runner/engine/el/InjectVariableELResolver.java
@@ -165,16 +165,10 @@ public class InjectVariableELResolver extends ELResolver {
         return result;
     }
 
-    @SuppressWarnings("deprecation")
     private static String getInjectedVariableName(Parameter p) {
         InjectVariable iv = p.getAnnotation(InjectVariable.class);
         if (iv != null) {
             return iv.value();
-        } else {
-            com.walmartlabs.concord.common.InjectVariable iv2 = p.getAnnotation(com.walmartlabs.concord.common.InjectVariable.class);
-            if (iv2 != null) {
-                return iv2.value();
-            }
         }
         return null;
     }